Added boot.asm file
While this isn't related to multithreading per se, this is related to my ultimate goal with multithreading, which is kernel development. This is a basic bootloader reflecting my basic progress with kernel development; compile this under "nasm -f bin ./boot.asm -o ./boot.bin" and run with "qemu-system-x86_64 -hda ./boot.bin" to get a basic boot up in QEMU that displays "McKyle OS", the introductory name for my proposed kernel.